feat: google oauth 인가 서버에 auth code 보내고 GoogleTokenResponse 받아오기
- 모킹을 이용하여 외부 API 테스트.
- RestTemplate을 @Bean으로 등록하여 싱글톤으로 사용
- RestTemplate이 여러 번 호출된다면, Thread safe하고 무겁기에 매번 호출해서 쓰는 것보다 싱글톤으로 쓰는 것이 효율적임
(Once created (or built), the RestClient can be used safely by multiple threads. - spring docs)

+++ src/main/java/app/cardcapture/auth/google/dto/GoogleTokenResponseDto.java
@@ -13,5 +13,5 @@ public class GoogleTokenResponseDto {
     private String refreshToken;
     private String idToken;
     private String tokenType;
-    private Long expiresIn;
+    private int expiresIn;
 }
